requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.31.0
Pillow==10.4.0
gunicorn==21.2.0
python-dotenv==1.0.0